shuffled everything to the right place
